Add files via upload
Commit#2: We covered Control Flow this week. We highlighted that Control Flow has three main components: Selection, Iteration and Recursion. If you still remember, we gave some examples of Selection Control Flow as 'if' statements, 'if-else' statements, 'switch case', etc. Example of Iteration Control Flow are 'for' loop, 'for-each' loop, 'while' loop, 'do-while' loop, etc. Take the examples in this code commit as a standard or perhaps a rubric in Java coding so far.